requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jieba3k==0.35.1
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
newspaper3k==0.2.8
nltk==3.9.1
oauthlib==3.2.2
openai==1.55.0
pillow==11.0.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
s3transfer==0.10.4
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
